MailerGui/requirements.txt
lxml==4.4.2
pandas==0.25.3
PyQt5==5.14.1
PyQt5-sip==12.7.0
pyqt5-tools==5.13.0.1.5
PyQtWebEngine==5.14.0
PyQtWebKit==5.13.1
python-dotenv==0.10.5
sendgrid==6.1.1

